refactor(channels): the post-send reconciliation is a module callable outside a turn
makeSentMessageIdReconciler and its helpers move verbatim from
channel-reply-delivery.ts into outbound-post-reconciliation.ts and the
reconciler is exported. The reply path composes the same function into
its delivery callback as before; no behavior changes. This lets a
producer that persists a row and delivers it through a channel adapter
outside a turn record the acknowledged provider id the way replies do.

Part of LUM-3525.
